+++ integ/job_cancellation_test.go
@@ -2,6 +2,7 @@ package integ
 
 import (
 	"fmt"
+	corev1 "k8s.io/api/core/v1"
 	"time"
 
 	"github.com/lyft/flinkk8soperator/pkg/apis/app/v1beta1"
@@ -56,12 +57,52 @@ func WaitUpdateAndValidate(c *C, s *IntegSuite, name string, updateFn func(app *
 	return newApp
 }
 
+func (s *IntegSuite) TestVertexNotRunning(c *C) {
+	log.Info("Starting test TestVertexNotRunning")
+
+	config, err := s.Util.ReadFlinkApplication("test_app.yaml")
+	c.Assert(err, IsNil, Commentf("Failed to read test app yaml"))
+
+	config.Name = "testvertexnotrunning"
+	config.ObjectMeta.Labels["integTest"] = "test_vertex_not_running"
+	envVar := corev1.EnvVar{
+		Name:  "EXTERNAL_CHECKPOINT",
+		Value: "1",
+	}
+
+	config.Spec.JobManagerConfig.EnvConfig.Env =
+		append(config.Spec.JobManagerConfig.EnvConfig.Env, envVar)
+	config.Spec.TaskManagerConfig.EnvConfig.Env =
+		append(config.Spec.TaskManagerConfig.EnvConfig.Env, envVar)
+	config.Spec.ProgramArgs = "--job_name testapp, --mode realtime"
+	c.Assert(s.Util.CreateFlinkApplication(config), IsNil,
+		Commentf("Failed to create flink application"))
+
+	log.Info("Application Created")
+
+	// wait for it to be running
+	c.Assert(s.Util.WaitForPhase(config.Name, v1beta1.FlinkApplicationDeployFailed), IsNil)
+
+	// delete the application
+	c.Assert(s.Util.FlinkApps().Delete(config.Name, &v1.DeleteOptions{}), IsNil)
+	for {
+		pods, err := s.Util.KubeClient.CoreV1().Pods(s.Util.Namespace.Name).
+			List(v1.ListOptions{LabelSelector: "integTest=test_recovery"})
+		c.Assert(err, IsNil)
+		if len(pods.Items) == 0 {
+			break
+		}
+	}
+	log.Info("All pods torn down")
+	log.Info("Completed test TestRecovery")
+}
+
 // tests the workflow of job cancellation without savepoint
 func (s *IntegSuite) TestJobCancellationWithoutSavepoint(c *C) {
 	log.Info("Starting test TestJobCancellationWithoutSavepoint")
 	testName := "cancelsuccess"
 	const finalizer = "simple.finalizers.test.com"
-
+	c.Skip("Skipping due to memory constraints in CI")
 	// start a simple app
 	config, err := s.Util.ReadFlinkApplication("test_app.yaml")
 	c.Assert(err, IsNil, Commentf("Failed to read test app yaml"))
@@ -138,7 +179,7 @@ func (s *IntegSuite) TestJobCancellationWithoutSavepoint(c *C) {
 // here, the new submitted job starts without a savepoint.
 func (s *IntegSuite) TestCancelledJobWithoutSavepoint(c *C) {
 	log.Info("Starting test TestCancelledJobWithoutSavepoint")
-
+	c.Skip("Skipping due to memory constraints in CI")
 	testName := "invalidcancel"
 	config, err := s.Util.ReadFlinkApplication("test_app.yaml")
 	c.Assert(err, IsNil, Commentf("Failed to read test app yaml"))
@@ -214,7 +255,7 @@ func (s *IntegSuite) TestCancelledJobWithoutSavepoint(c *C) {
 // tests the recovery workflow of the job when savepoint is disabled.
 func (s *IntegSuite) TestJobRecoveryWithoutSavepoint(c *C) {
 	log.Info("Starting test TestJobRecoveryWithoutSavepoint")
-
+	c.Skip("Skipping due to memory constraints in CI")
 	const finalizer = "simple.finalizers.test.com"
 	const testName = "cancelrecovery"
